requirements.txt
breadability==0.1.20
bson==0.5.8
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
Django==2.2
django-mongodb-engine-py3==0.6.0.1
djangorestframework==3.9.2
djangotoolbox==1.8.0
docopt==0.6.2
fuzzywuzzy==0.17.0
geographiclib==1.49
geopy==1.19.0
idna==2.8
kiwisolver==1.0.1
langdetect==1.0.7
lxml==4.3.3
matplotlib==3.0.3
Metaphone==0.6
nltk==3.4.1
numpy==1.16.3
pandas==0.24.2
pycountry==18.12.8
pygraphviz==1.5
pymongo==3.7.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
rake-nltk==1.0.4
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
sumy==0.7.0
urllib3==1.24.2
